updtate to 1.120.1xd
By the rivers of Babylon, there we sat down, yea, we wept, when we remembered Zion.

Remember, O Lord, The Children of Edom in the day of Jerusalem who said, "Raze it, raze it, even to the foundation."

O daughter of Babylon, who art to be destroyed. How happy shall he be, that rewardeth thee as thou hast served us.

Happy shall he be, that taketh and dasheth thy little ones against the stones.

Do you know what it means?